requirements.txt
# Core Web Framework
flask==3.0.0
flask-talisman==1.1.0
flask-limiter==3.5.0
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
gunicorn==21.2.0
eventlet==0.36.1

# Database
psycopg2-binary==2.9.9

# Kafka
kafka-python-ng>=2.2.3

# Environment Configuration
python-dotenv==1.0.0

# PDF Processing
PyPDF2>=3.0.0

# AI/LLM Integration
groq>=0.4.0

# HTTP Requests
requests==2.31.0

# ML and Data Analysis
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
joblib==1.3.2

# Deep Learning (LSTM)
tensorflow>=2.16.1

# PDF Generation (for reports)
reportlab==4.0.7
markdown==3.5.1

# OpenAI (legacy support)
openai==1.3.0
